Renderer
Factors out rendering a menu component from the component itself using the
visitor design pattern. Clients are required to subclass
`MenuComponentRenderer` and implement the rendering methods for each concrete
`MenuComponent` type.

This also generalises the subclass specific implementation of changing some
other state of a `MenuComponent`. When an item is selected, the `has_focus` can
be set to `true` so that `next` and `prev` can perform an action other than
iterating through the menu. For example, this allows `NumericMenuItem` to
increase/decrease its current value.